+++ src/main/java/com/lgcns/studify_be/comment/repository/CommentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.lgcns.studify_be.comment.domain.entity.CommentEntity;
 
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 
-    public List<CommentEntity> findByPost_PostId(Integer postId);
+    public List<CommentEntity> findByPost_PostId(Long postId);
 }
